requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
Flask-DebugToolbar==0.13.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.0
SQLAlchemy==2.0.19
alembic==1.12.1
WTForms==3.0.1
email-validator==2.0.0.post2
Werkzeug==2.3.6
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.6
backoff==2.2.1
Pillow==10.0.0
Flask-Uploads==0.2.1
Flask-Mail==0.9.1
